server-side/requirements.txt
flask
python-dotenv
langchain
langchain-google-genai
pandas
numpy
pymongo
flask-sqlalchemy
flask-bcrypt
flask-session
flask-cors
flask-migrate
deep-translator
langdetect
weasyprint
tavily-python
lingua-language-detector
reportlab
sentence-transformers
gtts
pytz
google-search-results
faiss-cpu
pypdf
google-generativeai
langchain-community
beautifulsoup4
PyMuPDF
Pillow
pykka
flask[async]
pypandoc-binary
pymongo
python-pptx